Issues/84 junit5 (#86)
* Update Bootstrap to 5.3.3 and remove JQuery #82
* test properties, fix SQL create table statements for H2 #84
* user is a reserved word in H2 -> quote in SQL #84
* use GreenMail as Mailserver during tests #84
* first Unit for create Domain #84
* first test for domains
* use Test-Profile - allows reuse of default application.properties #84
* H2 does not support table aliases in JOINs #84
    https://github.com/h2database/h2database/issues/2581
    Workaround: Use the complete table name
* start rewrite of PurlRestAPITest #84
* let retrieveUserFromRequest() return an Optional<User> #85
* fix visibility of Test classes and methods (Sonacloud) #84

+++ src/main/java/de/uni/rostock/ub/purl_server/controller/admin/AdminDomainController.java
@@ -285,10 +285,18 @@ public class AdminDomainController {
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @PostMapping(path = "/admin/manager/domain/delete")
     public String deleteDomain(@ModelAttribute Domain domain, HttpServletRequest request, Model model) {
-        domainDAO.retrieveDomain(domain.getId()).ifPresent(d -> {
-            domainDAO.deleteDomain(domain);
-            model.addAttribute(MODEL_ATTRIBUTE_DELETED, true);
-        });
+        Optional<User> u = purlAccess.retrieveCurrentUser();
+        Optional<Domain> d = domainDAO.retrieveDomain(domain.getId());
+        if (u.isPresent() && d.isPresent()) {
+            if (purlAccess.canModifyPurl(d.get(), u.get())) {
+                domainDAO.deleteDomain(domain);
+                model.addAttribute(MODEL_ATTRIBUTE_DELETED, true);
+            } else {
+                List<String> errorList = new ArrayList<>();
+                errorList.add(messages.getMessage("purl_server.error.domain.delete", null, Locale.getDefault()));
+                model.addAttribute(MODEL_ATTRIBUTE_ERRORS, errorList);
+            }
+        }
         model.addAttribute(MODEL_ATTRIBUTE_DOMAIN, domainDAO.retrieveDomain(domain.getId()).get());
         return MODEL_VIEW_DOMAINDELETE;
     }
